@@ -20,16 +20,14 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
-        "symfony/symfony": "~2.8|~3.0|^4.0",
-        "twig/twig": ">=1.5.0",
-        "doctrine/orm": "~2.4,>=2.4.5",
-        "doctrine/mongodb-odm": "^1.1.5"
+        "php": ">=7.3",
+        "symfony/symfony": "^4.0|^5.0",
+        "twig/twig": "^2.6|^3.0",
+        "doctrine/orm": "^2.6"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.7",
-        "friendsofphp/php-cs-fixer": "^2.0",
-        "satooshi/php-coveralls": "^1.0"
+        "phpunit/phpunit": "^7.0",
+        "friendsofphp/php-cs-fixer": "^2.0"
     },
     "suggest": {
         "ext-intl": "Translate the grid",
@@ -43,8 +41,5 @@
         "branch-alias": {
             "dev-master": "3.1-dev"
         }
-    },
-    "provide": {
-        "ext-mongo": "1.5"
     }
 }
