@@ -13,22 +13,20 @@
 namespace APY\DataGridBundle\Twig;
 
 use APY\DataGridBundle\Grid\Grid;
-use Pagerfanta\Adapter\NullAdapter;
-use Pagerfanta\Pagerfanta;
 use Symfony\Component\Routing\RouterInterface;
-use Twig_Environment;
-use Twig_Extension;
-use Twig_Extension_GlobalsInterface;
-use Twig_SimpleFilter;
-use Twig_SimpleFunction;
-use Twig_Template;
+use Twig\Environment;
+use Twig\Extension\AbstractExtension;
+use Twig\Extension\GlobalsInterface;
+use Twig\Template;
+use Twig\TwigFilter;
+use Twig\TwigFunction;
 
-class DataGridExtension extends Twig_Extension implements Twig_Extension_GlobalsInterface
+class DataGridExtension extends AbstractExtension implements GlobalsInterface
 {
     const DEFAULT_TEMPLATE = 'APYDataGridBundle::blocks.html.twig';
 
     /**
-     * @var Twig_Template[]
+     * @var Template[]
      */
     protected $templates = [];
 
@@ -53,103 +51,71 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     protected $params = [];
 
     /**
-     * @var array
-     */
-    protected $pagerFantaDefs;
-
-    /**
      * @var string
      */
     protected $defaultTemplate;
 
-    /**
-     * @param RouterInterface $router
-     * @param string          $defaultTemplate
-     */
-    public function __construct($router, $defaultTemplate)
+    public function __construct(RouterInterface $router, $defaultTemplate)
     {
         $this->router = $router;
         $this->defaultTemplate = $defaultTemplate;
     }
 
-    /**
-     * @param array $def
-     */
-    public function setPagerFanta(array $def)
-    {
-        $this->pagerFantaDefs = $def;
-    }
-
     public function getFilters()
     {
         return [
-            new Twig_SimpleFilter('data_grid_boolean_column_value_checker', [$this, 'booleanChecker']),
-            new Twig_SimpleFilter('data_grid_boolean_alt_value', [$this, 'booleanAltValueGetter']),
+            new TwigFilter('data_grid_boolean_column_value_checker', [$this, 'booleanChecker']),
+            new TwigFilter('data_grid_boolean_alt_value', [$this, 'booleanAltValueGetter']),
         ];
     }
 
-    /**
-     * @return array
-     */
-    public function getGlobals()
+    public function getGlobals(): array
     {
         return [
-            'grid'           => null,
-            'column'         => null,
-            'row'            => null,
-            'value'          => null,
+            'grid' => null,
+            'column' => null,
+            'row' => null,
+            'value' => null,
             'submitOnChange' => null,
-            'withjs'         => true,
-            'pagerfanta'     => false,
-            'op'             => 'eq',
+            'withjs' => true,
+            'op' => 'eq',
         ];
     }
 
-    /**
-     * Returns a list of functions to add to the existing list.
-     *
-     * @return array An array of functions
-     */
     public function getFunctions()
     {
         return [
-            new Twig_SimpleFunction('grid', [$this, 'getGrid'], [
+            new TwigFunction('grid', [$this, 'getGrid'], [
                 'needs_environment' => true,
-                'is_safe'           => ['html'],
-            ]),
-            new Twig_SimpleFunction('grid_html', [$this, 'getGridHtml'], [
-                'needs_environment' => true,
-                'is_safe'           => ['html'],
-            ]),
-            new Twig_SimpleFunction('grid_url', [$this, 'getGridUrl'], [
                 'is_safe' => ['html'],
             ]),
-            new Twig_SimpleFunction('grid_filter', [$this, 'getGridFilter'], [
+            new TwigFunction('grid_html', [$this, 'getGridHtml'], [
                 'needs_environment' => true,
-                'is_safe'           => ['html'],
-            ]),
-            new Twig_SimpleFunction('grid_column_operator', [$this, 'getGridColumnOperator'], [
-                'needs_environment' => true,
-                'is_safe'           => ['html'],
-            ]),
-            new Twig_SimpleFunction('grid_cell', [$this, 'getGridCell'], [
-                'needs_environment' => true,
-                'is_safe'           => ['html'],
-            ]),
-            new Twig_SimpleFunction('grid_search', [$this, 'getGridSearch'], [
-                'needs_environment' => true,
-                'is_safe'           => ['html'],
-            ]),
-            new Twig_SimpleFunction('grid_pager', [$this, 'getGridPager'], [
-                'needs_environment' => true,
-                'is_safe'           => ['html'],
-            ]),
-            new Twig_SimpleFunction('grid_pagerfanta', [$this, 'getPagerfanta'], [
                 'is_safe' => ['html'],
             ]),
-            new Twig_SimpleFunction('grid_*', [$this, 'getGrid_'], [
+            new TwigFunction('grid_url', [$this, 'getGridUrl'], [
+                'is_safe' => ['html'],
+            ]),
+            new TwigFunction('grid_filter', [$this, 'getGridFilter'], [
                 'needs_environment' => true,
-                'is_safe'           => ['html'],
+                'is_safe' => ['html'],
+            ]),
+            new TwigFunction('grid_column_operator', [$this, 'getGridColumnOperator'], [
+                'needs_environment' => true,
+                'is_safe' => ['html'],
+            ]),
+            new TwigFunction('grid_cell', [$this, 'getGridCell'], [
+                'needs_environment' => true,
+                'is_safe' => ['html'],
+            ]),
+            new TwigFunction('grid_search', [$this, 'getGridSearch'], [
+                'needs_environment' => true,
+                'is_safe' => ['html'],
+            ]),
+            new TwigFunction('grid_pager', [$this, 'getGridPager']),
+            new TwigFunction('grid_*', [$this, 'getGrid_'], [
+                'needs_environment' => true,
+                'is_safe' => ['html'],
             ]),
         ];
     }
@@ -172,7 +138,7 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     /**
      * Render grid block.
      *
-     * @param Twig_Environment              $environment
+     * @param Environment                   $environment
      * @param \APY\DataGridBundle\Grid\Grid $grid
      * @param string                        $theme
      * @param string                        $id
@@ -181,7 +147,7 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
      *
      * @return string
      */
-    public function getGrid(Twig_Environment $environment, $grid, $theme = null, $id = '', array $params = [], $withjs = true)
+    public function getGrid(Environment $environment, $grid, $theme = null, $id = '', array $params = [], $withjs = true)
     {
         $this->initGrid($grid, $theme, $id, $params);
 
@@ -194,7 +160,7 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     /**
      * Render grid block (html only).
      *
-     * @param Twig_Environment              $environment
+     * @param Environment                   $environment
      * @param \APY\DataGridBundle\Grid\Grid $grid
      * @param string                        $theme
      * @param string                        $id
@@ -202,56 +168,50 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
      *
      * @return string
      */
-    public function getGridHtml(Twig_Environment $environment, $grid, $theme = null, $id = '', array $params = [])
+    public function getGridHtml(Environment $environment, $grid, $theme = null, $id = '', array $params = [])
     {
         return $this->getGrid($environment, $grid, $theme, $id, $params, false);
     }
 
     /**
-     * @param Twig_Environment $environment
-     * @param string           $name
-     * @param unknown          $grid
+     * @param Environment $environment
+     * @param string      $name
+     * @param mixed       $grid
      *
      * @return string
      */
-    public function getGrid_(Twig_Environment $environment, $name, $grid)
+    public function getGrid_(Environment $environment, $name, $grid)
     {
         return $this->renderBlock($environment, 'grid_' . $name, ['grid' => $grid]);
     }
 
-    /**
-     * @param Twig_Environment $environment
-     * @param unknown          $grid
-     *
-     * @return string
-     */
-    public function getGridPager(Twig_Environment $environment, $grid)
+    public function getGridPager()
     {
-        return $this->renderBlock($environment, 'grid_pager', ['grid' => $grid, 'pagerfanta' => $this->pagerFantaDefs['enable']]);
+        return '';
     }
 
     /**
      * Cell Drawing override.
      *
-     * @param Twig_Environment                       $environment
+     * @param Environment                            $environment
      * @param \APY\DataGridBundle\Grid\Column\Column $column
      * @param \APY\DataGridBundle\Grid\Row           $row
      * @param \APY\DataGridBundle\Grid\Grid          $grid
      *
      * @return string
      */
-    public function getGridCell(Twig_Environment $environment, $column, $row, $grid)
+    public function getGridCell(Environment $environment, $column, $row, $grid)
     {
         $value = $column->renderCell($row->getField($column->getId()), $row, $this->router);
 
         $id = $this->names[$grid->getHash()];
 
         if (($id != '' && ($this->hasBlock($environment, $block = 'grid_' . $id . '_column_' . $column->getRenderBlockId() . '_cell')
-                    || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_' . $column->getType() . '_cell')
-                    || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_' . $column->getParentType() . '_cell')
-                    || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_id_' . $column->getRenderBlockId() . '_cell')
-                    || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_type_' . $column->getType() . '_cell')
-                    || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_type_' . $column->getParentType() . '_cell')))
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_' . $column->getType() . '_cell')
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_' . $column->getParentType() . '_cell')
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_id_' . $column->getRenderBlockId() . '_cell')
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_type_' . $column->getType() . '_cell')
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_type_' . $column->getParentType() . '_cell')))
             || $this->hasBlock($environment, $block = 'grid_column_' . $column->getRenderBlockId() . '_cell')
             || $this->hasBlock($environment, $block = 'grid_column_' . $column->getType() . '_cell')
             || $this->hasBlock($environment, $block = 'grid_column_' . $column->getParentType() . '_cell')
@@ -268,21 +228,21 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     /**
      * Filter Drawing override.
      *
-     * @param Twig_Environment                       $environment
+     * @param Environment                            $environment
      * @param \APY\DataGridBundle\Grid\Column\Column $column
      * @param \APY\DataGridBundle\Grid\Grid          $grid
      *
      * @return string
      */
-    public function getGridFilter(Twig_Environment $environment, $column, $grid, $submitOnChange = true)
+    public function getGridFilter(Environment $environment, $column, $grid, $submitOnChange = true)
     {
         $id = $this->names[$grid->getHash()];
 
         if (($id != '' && ($this->hasBlock($environment, $block = 'grid_' . $id . '_column_' . $column->getRenderBlockId() . '_filter')
-                    || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_id_' . $column->getRenderBlockId() . '_filter')
-                    || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_type_' . $column->getType() . '_filter')
-                    || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_type_' . $column->getParentType() . '_filter'))
-                || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_filter_type_' . $column->getFilterType()))
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_id_' . $column->getRenderBlockId() . '_filter')
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_type_' . $column->getType() . '_filter')
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_type_' . $column->getParentType() . '_filter'))
+            || $this->hasBlock($environment, $block = 'grid_' . $id . '_column_filter_type_' . $column->getFilterType()))
             || $this->hasBlock($environment, $block = 'grid_column_' . $column->getRenderBlockId() . '_filter')
             || $this->hasBlock($environment, $block = 'grid_column_id_' . $column->getRenderBlockId() . '_filter')
             || $this->hasBlock($environment, $block = 'grid_column_type_' . $column->getType() . '_filter')
@@ -298,14 +258,14 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     /**
      * Column Operator Drawing override.
      *
-     * @param Twig_Environment                       $environment
+     * @param Environment                            $environment
      * @param \APY\DataGridBundle\Grid\Column\Column $column
      * @param \APY\DataGridBundle\Grid\Grid          $grid
      * @param bool                                   $submitOnChange
      *
      * @return string
      */
-    public function getGridColumnOperator(Twig_Environment $environment, $column, $grid, $operator, $submitOnChange = true)
+    public function getGridColumnOperator(Environment $environment, $column, $grid, $operator, $submitOnChange = true)
     {
         return $this->renderBlock($environment, 'grid_column_operator', ['grid' => $grid, 'column' => $column, 'submitOnChange' => $submitOnChange, 'op' => $operator]);
     }
@@ -340,41 +300,19 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     }
 
     /**
-     * @param Twig_Environment $environment
-     * @param unknown          $grid
-     * @param unknown          $theme
-     * @param string           $id
-     * @param array            $params
+     * @param Environment $environment
+     * @param mixed       $grid
+     * @param mixed       $theme
+     * @param string      $id
+     * @param array       $params
      *
      * @return string
      */
-    public function getGridSearch(\Twig_Environment $environment, $grid, $theme = null, $id = '', array $params = [])
+    public function getGridSearch(Environment $environment, $grid, $theme = null, $id = '', array $params = [])
     {
         $this->initGrid($grid, $theme, $id, $params);
 
         return $this->renderBlock($environment, 'grid_search', ['grid' => $grid]);
-    }
-
-    /**
-     * @param unknown $grid
-     */
-    public function getPagerfanta($grid)
-    {
-        $adapter = new NullAdapter($grid->getTotalCount());
-
-        $pagerfanta = new Pagerfanta($adapter);
-        $pagerfanta->setMaxPerPage($grid->getLimit());
-        $pagerfanta->setCurrentPage($grid->getPage() + 1);
-
-        $url = $this->getGridUrl('page', $grid, '');
-        $routeGenerator = function ($page) use ($url) {
-            return sprintf('%s%d', $url, $page - 1);
-        };
-
-        $view = new $this->pagerFantaDefs['view_class']();
-        $html = $view->render($pagerfanta, $routeGenerator, $this->pagerFantaDefs['options']);
-
-        return $html;
     }
 
     /**
@@ -386,7 +324,7 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     {
         if (is_string($value)) {
             $value = strtolower($value);
-            if ($value == 'true' or $value == 'si') {
+            if ($value == 'true' || $value == 'si') {
                 return 'true';
             }
 
@@ -415,15 +353,15 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     /**
      * Render block.
      *
-     * @param Twig_Environment $environment
-     * @param string           $name
-     * @param array            $parameters
+     * @param Environment $environment
+     * @param string      $name
+     * @param array       $parameters
      *
      * @throws \InvalidArgumentException If the block could not be found
      *
      * @return string
      */
-    protected function renderBlock(Twig_Environment $environment, $name, $parameters)
+    protected function renderBlock(Environment $environment, $name, $parameters)
     {
         foreach ($this->getTemplates($environment) as $template) {
             if ($template->hasBlock($name, [])) {
@@ -437,15 +375,15 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     /**
      * Has block.
      *
-     * @param Twig_Environment $environment
-     * @param $name string
+     * @param Environment $environment
+     * @param mixed       $name
      *
      * @return bool
      */
-    protected function hasBlock(Twig_Environment $environment, $name)
+    protected function hasBlock(Environment $environment, $name)
     {
         foreach ($this->getTemplates($environment) as $template) {
-            /** @var $template Twig_Template */
+            /** @var Template $template */
             if ($template->hasBlock($name, [])) {
                 return true;
             }
@@ -457,21 +395,21 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     /**
      * Template Loader.
      *
-     * @param Twig_Environment $environment
+     * @param Environment $environment
      *
      * @throws \Exception
      *
-     * @return Twig_Template[]
+     * @return Template[]
      */
-    protected function getTemplates(Twig_Environment $environment)
+    protected function getTemplates(Environment $environment)
     {
         if (empty($this->templates)) {
-            if ($this->theme instanceof Twig_Template) {
+            if ($this->theme instanceof Template) {
                 $this->templates[] = $this->theme;
                 $this->templates[] = $environment->loadTemplate($this->defaultTemplate);
             } elseif (is_string($this->theme)) {
                 $this->templates = $this->getTemplatesFromString($environment, $this->theme);
-            } elseif ($this->theme === null) {
+            } elseif (null === $this->theme) {
                 $this->templates = $this->getTemplatesFromString($environment, $this->defaultTemplate);
             } else {
                 throw new \Exception('Unable to load template');
@@ -482,17 +420,17 @@ class DataGridExtension extends Twig_Extension implements Twig_Extension_Globals
     }
 
     /**
-     * @param Twig_Environment $environment
-     * @param unknown          $theme
+     * @param Environment $environment
+     * @param mixed       $theme
      *
-     * @return array|Twig_Template[]
+     * @return Template[]
      */
-    protected function getTemplatesFromString(Twig_Environment $environment, $theme)
+    protected function getTemplatesFromString(Environment $environment, $theme)
     {
         $this->templates = [];
 
         $template = $environment->loadTemplate($theme);
-        while ($template instanceof \Twig_Template) {
+        while ($template instanceof Template) {
             $this->templates[] = $template;
             $template = $template->getParent([]);
         }
